requirements.txt
torch~=2.2.2+cu118
pandas~=2.2.2
torchvision~=0.17.2
pillow~=10.3.0
tqdm~=4.66.2
PyQt5~=5.15.10